19. 29.01.23/requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pytz==2022.7
yarl==1.8.2
APScheduler==3.6.3
cachetools==4.2.2
isOdd==0.1.2
python-telegram-bot==13.14
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
prettytable==3.6.0
